change session_id to 64-bit value so the userspace value matches the database value.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@30790 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ http-casing/api/com/untangle/node/http/RequestLine.java
@@ -124,12 +124,12 @@ public class RequestLine implements Serializable
      * @return the PipelineEndpoints.
      */
     @Column(name="session_id", nullable=false)
-    public Integer getSessionId()
+    public Long getSessionId()
     {
         return pipelineEndpoints.getSessionId();
     }
 
-    public void setSessionId( Integer sessionId )
+    public void setSessionId( Long sessionId )
     {
         this.pipelineEndpoints.setSessionId(sessionId);
     }
