+++ jnetcap/impl/com/untangle/jnetcap/NetcapHook.java
@@ -1,8 +1,10 @@
-/* $HeadURL$ */
+/**
+ * $Id$
+ */
 package com.untangle.jnetcap;
 
 public interface NetcapHook
 {
     /* This the callback that will be called for the UDP/TCP hooks */
-    public void event( int sessionId );
+    public void event( long sessionId );
 }
